updated the sdk according to code review feedback
