+++ sdk/communication/communication-call-automation/src/index.ts
@@ -5,8 +5,6 @@ export * from "./callConnection";
 export * from "./callMedia";
 export * from "./callRecording";
 export * from "./callAutomationEventParser";
-export * from "./models/constants";
-export * from "./models/logger";
 export * from "./models/models";
 export * from "./models/options";
 export * from "./models/responses";
