+++ sdk/communication/communication-call-automation/src/models/constants.ts
@@ -1,4 +1,4 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-export const SDK_VERSION: string = "1.0.0-beta.2";
+export const SDK_VERSION: string = "1.0.0-beta.1";
